17 de octubre 2012 mañana, pruebas del adc con el capturador de freciencias

+++ accelerometro.h
@@ -123,9 +123,9 @@ typedef union
   } reg;
   struct
   {
-    int x;
-    int y;
-    int z;
+    int16 x;
+    int16 y;
+    int16 z;
   } value;
 } xyz_union;
